+++ source/backend/discovery/src/lib/config.js
@@ -8,6 +8,7 @@ const config = {
     cluster: process.env.CLUSTER,
     crossAccountDiscovery: process.env.CROSS_ACCOUNT_DISCOVERY,
     isUsingOrganizations: process.env.CROSS_ACCOUNT_DISCOVERY === AWS_ORGANIZATIONS,
+    organizationUnitId: process.env.ORGANIZATION_UNIT_ID,
     configAggregator: process.env.CONFIG_AGGREGATOR,
     graphgQlUrl: process.env.GRAPHQL_API_URL,
     rootAccountRole: process.env.DISCOVERY_ROLE,