+++ source/frontend/src/components/Explore/Resources/Types/TypeOverview/AccountMultiSelect.js
@@ -4,17 +4,25 @@
 import React from "react";
 import {Button, FormField, Multiselect} from "@cloudscape-design/components";
 import * as R from "ramda";
-import {useResourcesRegionMetadata} from "../../../../Hooks/useResourcesMetadata";
+import {
+    useResourcesMetadata,
+    useResourcesRegionMetadata
+} from "../../../../Hooks/useResourcesMetadata";
 import {useDeepCompareEffect} from "react-use";
 
+
 const AccountMultiSelect = ({selected=[], onChange=() => ({}), onOptionsChange=() => ({}), disabled=false}) => {
-  const { data: accounts=[], status } = useResourcesRegionMetadata();
+  const { data: resources= {accounts: []} } = useResourcesMetadata();
+  const accountsFilter = resources.accounts.map(({accountId}) => ({accountId}));
+  const { data: accountsRegionMetadata=[] } = useResourcesRegionMetadata(accountsFilter, {
+      batchSize: 50
+  });
 
   useDeepCompareEffect(() => {
-    onOptionsChange(accounts.map(i => i.accountId));
-  }, [accounts, onOptionsChange])
+    onOptionsChange(accountsRegionMetadata.map(i => i.accountId));
+  }, [accountsRegionMetadata, onOptionsChange])
 
-  const options = accounts.map(i => ({
+  const options = accountsRegionMetadata.map(i => ({
     label: i.accountId,
     value: i.accountId,
     tags:[
@@ -33,7 +41,6 @@ const AccountMultiSelect = ({selected=[], onChange=() => ({}), onOptionsChange=(
         filteringType="auto"
         disabled={disabled}
         options={options}
-        status={status}
         selectedOptions={options.filter(i => selected.indexOf(i.value) !== -1)}
         onChange={e => onChange(e.detail.selectedOptions.map(i => i.value))}
         placeholder="Choose accounts to filter by"